test(store): attempt to fix and clean code
